refactor(model): rename domain to model
