+++ src/main/kotlin/racingcar/model/Car.kt
@@ -1,4 +1,4 @@
-package racingcar.domain
+package racingcar.model
 
 data class Car(val name: String, var position: Int = 0) {
     fun move() {
